+++ Tests/Security/Authentication/Firewall/JWTListenerTest.php
@@ -9,6 +9,8 @@ use Lexik\Bundle\JWTAuthenticationBundle\Security\Firewall\JWTListener;
 /**
  * JWTListenerTest.
  *
+ * @group legacy
+ *
  * @author Nicolas Cabot <n.cabot@lexik.fr>
  * @author Robin Chalas <robin.chalas@gmail.com>
  */